RuItUnion.FeedbackBot.Data.Old refactoring
